[sprinkler][dfplayer][max6956][rf_bridge] Fix cg.templatable type mismatches
The type argument to cg.templatable() must match the C++
TEMPLATABLE_VALUE type. When a lambda is used, this type becomes the
lambda return type — a mismatch generates incorrect C++.

sprinkler: cg.float_ → cg.uint32 for set_divider and set_repeat
dfplayer: float → cg.uint16/cg.bool_/cg.uint8 for file/loop/volume
max6956: float → cg.uint8/enum for brightness_global/brightness_mode
rf_bridge: cg.uint16 → cg.uint8 for length/protocol

+++ esphome/components/rf_bridge/__init__.py
@@ -185,9 +185,9 @@ RFBRIDGE_SEND_ADVANCED_CODE_SCHEMA = cv.Schema(
 async def rf_bridge_send_advanced_code_to_code(config, action_id, template_args, args):
     paren = await cg.get_variable(config[CONF_ID])
     var = cg.new_Pvariable(action_id, template_args, paren)
-    template_ = await cg.templatable(config[CONF_LENGTH], args, cg.uint16)
+    template_ = await cg.templatable(config[CONF_LENGTH], args, cg.uint8)
     cg.add(var.set_length(template_))
-    template_ = await cg.templatable(config[CONF_PROTOCOL], args, cg.uint16)
+    template_ = await cg.templatable(config[CONF_PROTOCOL], args, cg.uint8)
     cg.add(var.set_protocol(template_))
     template_ = await cg.templatable(config[CONF_CODE], args, cg.std_string)
     cg.add(var.set_code(template_))
